pages/api/spotify_api/image_requirements.txt
google-generativeai==0.3.0
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
